requirements.txt
fastapi
uvicorn
openai
python-dotenv
sse-starlette
jinja2
langchain>=0.1.0
langchain-openai
langchain-google-genai
langchain-community>=0.0.20
python-multipart
python-magic
itsdangerous
# Document processing dependencies
pymupdf>=1.22.0
pypdf>=4.0.0
unstructured[pdf]>=0.10.0
# Additional text processing support
chardet>=5.0.0
# Chunking dependencies
langchain-text-splitters>=0.0.1
tiktoken>=0.5.0